folder system demo

+++ src/pages/manager/fileTable.js
@@ -1,38 +1,60 @@
 import React, {Component} from 'react'
-import { Table, Divider, Button } from 'antd'
+import { Table, Divider, Button, Tag } from 'antd'
+import {getRenderTree} from './transTree'
 function columnWrapper (self) {
   return [
     {
-      title: 'title',
-      dataIndex: 'title',
-      key: 'title'
+      title: 'Name',
+      dataIndex: 'name',
+      key: 'name'
     },
     {
-      title: 'type',
+      title: 'Type',
+      key: 'type',
       dataIndex: 'type',
-      key: 'type'
-    },
-    {
-      title: 'createDate',
-      dataIndex: 'createDate',
-      key: 'createDate'
+      render: (text, record) => {
+        let color = 'geekblue'
+        if (text === 'folder') {
+          color = 'green'
+        }
+        return (<Tag color={color} key={text}>{text.toUpperCase()}</Tag>)
+      }
     },
     {
       title: 'Action',
       key: 'action',
-      render: (text, record) => (
-        <span>
-          <a onClick={() => {self.editCol(record)}}>Edit</a>
-          <Divider type="vertical" />
-          <a onClick={() => {self.deleteFile(record)}}>Delete</a>
-        </span>
-      )
+      render: (text, record) => {
+        if (record.type !== 'folder') {
+          return (
+            <span>
+              <a onClick={() => {self.editCol(record)}}>Edit</a>
+              <Divider type="vertical" />
+              <a onClick={() => {self.deleteFile(record)}}>Delete</a>
+            </span>
+          )
+        }
+      }
     }
   ]
 }
 class FileTable extends Component {
+  constructor (props) {
+    super()
+    const {dataSource} = props
+    this.state = {
+      renderTree: getRenderTree(dataSource)
+    }
+  }
+  UNSAFE_componentWillReceiveProps (props) {
+    const {dataSource} = props
+    this.setState({
+      renderTree: getRenderTree(dataSource)
+    })
+  }
   editCol = (record) => {
-    let index = this.props.dataSource.indexOf(record)
+    let index = this.props.dataSource.findIndex(item => {
+      return item.id === record.id
+    })
     this.props.dispatch({
       type: 'manager/getFile',
       payload: {
@@ -65,11 +87,18 @@ class FileTable extends Component {
     })
   }
   render () {
-    const {dataSource} = this.props
+    const {renderTree} = this.state
+    console.log(renderTree)
     return (
       <div>
         <Button type="primary" icon="upload" onClick={this.addFile}>Upload</Button>
-        <Table style={{marginTop: '1rem'}} columns={columnWrapper(this)} dataSource={dataSource} rowKey={(record) => record.id} />
+        <Button type="primary" icon="plus" style={{marginLeft: '10px'}}>Add Folder</Button>
+        <Table style={{marginTop: '1rem'}}
+          columns={columnWrapper(this)}
+          dataSource={renderTree.children}
+          // expandedRowRender={this.dfsRender}
+          // rowKey={(record, index) => {return record.name + index}}
+          />
       </div>
     )
   }
